added udp listener

+++ OpenPath/Classes/Core/RayPart/RSocket.h
@@ -75,9 +75,6 @@ method(RSocket *,  accept,        RSocket));
 
 // Main methods
 method(byte,  send,               RSocket),    const pointer buffer, size_t size);  // sends to receiver size bytes
-//extern
-//method(byte,  sendString,         RSocket),    const RCString *string);  // sends to receiver size bytes
-
-method(byte,  receive,            RSocket),    pointer buffer, size_t size);  // buffer must be pre allocated at least 1500 bytes, return 255 if fails, 1 if received some
+method(ssize_t,  receive,            RSocket),    pointer buffer, size_t size);  // buffer must be pre allocated at least 1500 bytes, return 255 if fails, 1 if received some
 
 #endif /*__R_SOCKET_H__*/
